requirements.txt
streamlit
PyMuPDF
google-generativeai
python-dotenv
reportlab
pillow
